# temandifa-backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TemanDifa API",
    "description": "API for TemanDifa - Assistive AI Application for the Visually Impaired",
    "contact": {
      "name": "TemanDifa Support",
      "email": "support@temandifa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
...
  "paths": {
...
    "/history": {
      "get": {
        "tags": ["History"],
        "summary": "Get user history",
        "description": "Retrieve paginated user activity history",
        "produces": ["application/json"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "type": "integer", "default": 1 },
          { "in": "query", "name": "limit", "type": "integer", "default": 20 }
        ],
        "responses": {
          "200": { "description": "History list with pagination" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["History"],
        "summary": "Create history entry",
        "description": "Save a new activity to history",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["feature_type", "result_text"],
              "properties": {
                "feature_type": {
                  "type": "string",
                  "enum": ["OBJECT", "OCR", "VOICE"]
                },
                "input_source": { "type": "string" },
                "result_text": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "History saved" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["History"],
        "summary": "Clear all history",
        "description": "Delete all history entries for the user",
        "produces": ["application/json"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "All history cleared" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/history/{id}": {
      "delete": {
        "tags": ["History"],
        "summary": "Delete history entry",
        "description": "Delete a specific history entry",
        "produces": ["application/json"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "History deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/cache/stats": {
      "get": {
        "tags": ["Cache"],
        "summary": "Get cache statistics",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Cache stats" } }
      }
    },
    "/cache": {
      "delete": {
        "tags": ["Cache"],
        "summary": "Clear all AI cache",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Cache cleared" } }
      }
    },
...
  }
}